Simplified module listings by putting a list of core modules into each package initialization script. The utility functions are augmented to recursively process these lists of core modules, and to skip function calls if the functions are absent. This allowed me to drop empty load_data() functions from a handful of modules as well. Splitting components definitions into standard and dynamic components was an essential part of this task because switch_mod needed to define standard components before other modules in a given list, and define dynamic components after all standard components are defined.
The calling conventions are backwards compatible so you can specify core project modules as either (..., 'project', ...) or as (..., 'project.build', 'project.dispatch', ...).

I updated the examples to take advantage of the new calling convention because it is more compact. I also tweaked copperplate0 to ignore local transmission & distribution to illustrate that module is optional and can be ignored or substituted with a different treatment of those components.