Test runner: Replace the shell script with a Python script
This is faster, because it starts up Python and loads the switch_mod
modules only once, instead of once per module.  This reduces the time
to run the tests from 10.9s to 1.7s on my laptop.

This also produces nicer output, because it reuses Python's unittest
framework, which prints a progress bar.