Test the objective function value (total cost) generated for the examples
switch_mod/utilities.py: Add the ability to export the value of the
objective function to total_cost.txt.

Check in expected values for the examples' total costs in the
"examples" directory.

Change examples_test.py to test for those expected values, with an
option ("--update") for updating them easily whenever changes to the
model require this in future.