Fuel_Costs_TP Expression speedup (#40)
The Fuel_Costs_TP Expression in the fuel_cost module is taking significant time to be constructed. 
For a medium sized system and time resolution (23 lz, 720 tp, 190 proj, 32 tx lines) the original code takes between 4.3 and 4.8 s on my machine to construct the Fuel_Costs_TP expression, while with this new formulation -provided by Matthias- that time is reduced to between 0.85 and 0.95 s. Outputs are identical, so everything is working as expected.