Adding a new module to model complex (or not so complex) hydroelectric systems.
It has a high degree of flexibility, allowing for modeling of most relevant operational phenomena.

Code can be made more efficient, but I've separated some sets and constraints on purpose so that
the module is more readable.

I pretend to upload an example sometime soon, but meanwhile you can check the module itself.

The boundary conditions are a pain, since the model will always tend to use up all available water.
There are a couple of commonly used conditions:
(i) Including a water cost for the volume that is stored in the final timepoint the simulation (a Future Cost Function)
(ii) Forcing each reservoir to reach the end of the simulation with a certain volume of water.

I hard coded in the latter, though with a period resolution. So, each reservoir will end each period with the inputted volume, at least.

More flexibilization needs to be added, so that people can modify the hydro system's topology within
the simulation horizon (now it's considered to be fixed) and to allow using different boundary conditions.

I tried documenting it as best as I could. If there is anything that is not clear or you find a bug, please
let me know and I'll fix it.