Minimum capacity build constraint (#63)
Implemented minimum capacity build constraint. This incorporates binary variables into the model.
Includes an example to show how minimum capacity build works. This is an extension of the discrete_build example. Also, fixed the README file for the latter.

The user may turn these constraints on or off by inputting or not the g_min_build_capacity parameter for each gentech.