A functional draft of spinning reserves based on the Hawaii module.
The spinning_reserves module creates four dynamic lists for upward and downward requirements and provisions that act in an additive manner, as well as a n-1 contigency list.

A few spinning requirement rules and contigency behaviors can be selectively enabled with configuration options (command line parameters/config file).

The Hawaii non-generator energy submodules (ev, dr, hydrogen, etc) self-register with the spinning reserve provision list if it is available. I started moving semi-documented Hawaii reserve requirements into separate Hawaii submodule, with the plan of putting it back as a function in spinning_reserves when it gets documentation and references.