Move save_results to core and post_solve on reload
It is sometimes necessary to reload models to run new post_solve code,
e.g., if new outputs are defined after time-consuming models have
already been run. Previously, --reload-prior-solution just reloaded
the model and dumped the user into an interactive prompt, with no easy
way to run module code.

This commit changes the behavior of --reload-prior-solution, so that it
now acts like an exact alternative to re-running the model, i.e., if the
user specifies --reload-prior-solution, we reload the solution, then run
the normal post-solve code, and don't drop to an interactive prompt
unless they also specify --interact.

Since this behavior is now central to the use of the model rather than
a quasi-reporting behavior, this commit also moves save_results() back
from switch_model.reporting to switch_model.solve.

This commit also adds a --no-save-solution flag to disable automatic
solution- saving, which is helpful for models that will be solved
repeatedly and don't need to be reloaded.