Patch Pyomo to accelerate reloading prior solution
Pyomo 5.1.1 (and probably other versions) is very slow to load prior
solutions because it does a full-component search for each component
name as it assigns the data. This creates a delay that is quadratic in
model size, so reloading solutions takes longer than solving the model
from scratch.

This code monkey-patches a few lines (!) of
pyomo.core.base.PyomoModel.ModelSolutions.add_solution to use Pyomo's
built-in caching system for component names. This makes
--load-prior-solution fast enough to use for practical work. This is a
pretty extreme approach, but there doesn't seem to be another way short
of writing our own model-reload code.

TODO: create a pull request for Pyomo to do this