Smooth ChargeEVs even when it is an Expression
The hawaii.ev_advanced module constructs ChargeEVs as a weighted sum of
several possible charging profiles. If this is squared and added to the
smoothing objective, it makes the model non-positive-definite, so cplex
won't solve it. This code instead smoothes a helper variable
ChargeEVsVar that is bound to ChargeEVs. This commit also refactors the
smoothing code to make it a little DRYer.