Minimize changes instead of squared values
hawaii.smooth_dispatch previously smoothed the hourly dispatch profile
for slack elements (storage, demand response, EVs) by minimizing the sum
of the squares of the hourly values of these elements. This requires a
quadratic solver and is difficult to setup with hawaii.ev_advanced,
which defines EV charging as a convex combination of other decision
variables. With this commit, we instead switch to minimizing the hour-
to-hour variation of these variables (actually the total upward
variation over the study period), which is linear, faster to solve, and
works about as well.