Add '--exact' option to hawaii.rps, to require exactly matching RPS target
This is useful for studying how much _extra_ it costs if you restrict 
renewables to the target instead of allowing a lower-cost system with 
more renewables. It can be used to graphically show the "sweet spot" 
where more renewables decrease costs up to a certain point and then 
begin to increase costs.