Skip carbon cap if not specified, instead of limiting to float('inf')
We have had cases where solvers took excessively long to solve when 
infinity was used as a bound on a constraint (maybe it gets passed as 
the maximum float and causes numeric problems?). This issue needs more 
investigation, and we may want to skip other constraints or bounds that 
use float('inf') as well.