Simplified and fixed bugs in local_td capacity and cost calculations
Pyomo under Python 3 does not support mixing data types in index keys 
for
variables (it sorts the index keys before sending the variable to the 
solver,
and that causes a comparison error on Python 3). We previously used 
"Legacy" as
a special index key for BuildLocalTD (representing capacity online prior 
to the
study), along with the study periods, which are ints. Since this is no 
longer
allowed, this commit simplifies the code to track existing local T&D 
capacity
separately from new T&D capacity.

These changes eliminate the need for the LOCAL_TD_BLD_YRS set, so I have 
removed
that. I also simplified the calculation of LocalTDCapacity, using a new
CURRENT_AND_PRIOR_PERIODS_FOR_PERIOD set.

omitted from the fixed cost calculations due to an error comparing each 
period
to prior periods ('Legacy' <= 2030 is False in Python 2, but was assumed 
to be
True in the code). This has been fixed now.

These changes have several side effects:

- local T&D fixed costs will (correctly) be higher than in Switch 2.0.3 
for all
  code that uses switch_model.transmission.local_td
- the LOCAL_TD_BLD_YRS set has been removed from the model
- the CURRENT_AND_PRIOR_PERIODS_FOR_PERIOD set is now defined in
  switch_model.timescales
- BuildLocalTD.tab in the outputs directory now has BuildLocalTD_index_1 
and
  BuildLocalTD_index_2 as the first two column headings instead of
  LOCAL_TD_BLD_YRS_1 and LOCAL_TD_BLD_YRS_2
- BuildLocalTD.tab in the outputs directory no longer includes a 
"Legacy" row
- solutions to models that use switch_model.transmission.local_td that 
were
  saved by Switch before version 2.0.4 cannot be opened in 2.0.4 or 
later