Make run_tests.py compatible with Python 3 (and 2)
Unlike Python 2, Python 3 only applies the `discover` command if the 
active module is `None`. This gives undefined argument errors when 
trying to run unit tests with the old code. Setting `module=None` in 
`unittest.TestProgram()` fixes this.

In addition, Python 3 passes the search pattern as a keyword arg to the 
custom testLoader, but Python 2 does not. To achieve compatibility, we 
just pass all kwargs through to the parent class.