Use conda to install dependencies, then pip to install Switch
It is easiest to install dependencies for Switch using the conda 
command, but Switch itself must currently be installed via the pip 
command.  Anaconda recommends minimizing use of pip in conda 
environments 
(https://www.anaconda.com/using-pip-in-a-conda-environment/). In 
particular, they recommend installing dependencies with conda first, 
then using pip for the packages that can't be installed with conda. We 
have also had problems using pip to install Switch and its dependencies 
under Anaconda in one pass. e.g., Pyomo depends on nose, but when pip is 
used to install nose in a conda environment on a Mac, pip attempts to 
write a man page to /usr/local/man, which is not generally 
user-writable, so the installation can fail. Or using pip to install 
rpy2 will fail if R is not already installed.

So, instead, we now advise users to install the dependencies with conda 
first, then use pip to install Switch. This is more verbose than 
previously, but should avoid some difficult installation problems, and 
requires only about as much thinking as the previous question of whether 
to use `pip install switch-model` or `pip install 
switch-model[advanced]`, etc.

This rewrite of the INSTALL file also offers the option of installing 
Switch from pypi (a "minimal" install) or installing in a local 
directory (a "complete" install), and explains the differences.

This rewrite also makes cbc more parallel to glpk and downplays the 
possible performance gain (I found cbc was about 10% slower than glpk 
for a couple of test models).

To support use of cbc, I also removed the default solver suffixes, since 
cbc doesn't support the 'slack' suffix.

We also now recommend that users install pint along with Pyomo, because 
Pyomo requires pint (their unit testing framework) when we run our unit 
tests. We may eventually want to group testfixtures and pint with git as 
things that people will want only if they are going to do the "complete" 
install (and run tests). Then we could remove testfixtures from 
install_requires in setup.py.