Add license file, tests and other information to distribution
We are required to distribute a copy of the license file with the code. 
Including license and tests here also makes it easier to include and use 
them in conda-forge. 

These extra files also enable users to download the source code and 
examples to an editable location without using git. Then they can finish 
installation via `pip install --editable .`