Fix .tab parsing errors in upgrade scripts and fix malformed .tab files
There has been an extra 1 value in the Hydro_RoR row of
upgrade_dat/hydro_simple/inputs/proj_build_costs.tab for a long time. In addition, the Hydro
and Hydro_RoR rows were delimited with spaces (often multiple) instead of tab characters.
These problems also occurred in the original examples/hydro_simple directory, which has been
repeatedly upgraded, creating more malformed files along the way. Spaces were also used in
some of the other example files, creating similar problems.

These were not caught previously because the following sequence happened:

Pyomo was happy to load the original version of these files (2.0.0b0), because it allows space
or tab delimiters. It also ignores extra fields on each row, so it just used the first four
fields (including the extra 1) and ignored the fifth field. So this example ran with the
wrong values for Hydro_RoR in the third and fourth columns until now.

The upgrade scripts for later versions read these files using pandas.read_csv with separator
set to '\t'. For all the rows delimited with spaces, this meant that the upgrade script read
the whole row into the first field and filled the other fields with nulls. Then when the
script wrote the file back, it preserved the spaces between the values that had been
pulled into the first field, and added on extra '.' values for the no-value fields.
Subsequent upgrade scripts ran fine, because they still read all the space-delimited stuff
into the first field, and read the '.' values into other fields. Switch was also happy to run
with these files at each step, because Pyomo used the spaces as field delimiters and ignored
the extra '.' values at the end of the rows.

However, when we change to .csv inputs, this no longer works. The upgrade script could read
these files in with '\t' delimiters, but then most of the values end up in the first column,
and they can't be read correctly by Pyomo when it reads these files expecting ',' delimiters.

The solution is as follows: (1) update the upgrade scripts to use r'\s+' (any whitespace) as
delimiters when reading .tab files (pre 2.0.5), so that they will then write
out correctly formatted, \t-delimited .tab files without extra columns. This fixes the
upgrade tests. (2) remove the extra "1" value from the Hydro_RoR rows and update the
total_cost.txt to reflect this (in examples and tests).