Implement --input-aliases option
The command-line argument for this was accidentally included in a
previous version; this implements the actual behavior.

Filename substitutions are specified as
--input-alias[es] file1.csv=file1.alternate.csv
[file2.csv=file2.alternate.csv, ...]

These are applied as a simple replacement on the filename at the end of
the path specified in the call to load_aug(). So usually the replacement
will occur within the --inputs-dir, but users can specify
'some_dir/file.csv' or '../some_other_dir/file.csv' to refer to files in
other directories relative to the normal location of the file.

Filename 'none' will be converted to an empty string and usually cause
the file to be ignored.

This enables use of alternative files to study sensitivities without
creating complete input directories for each permutation.