Switch from --verbose/--quiet to --log-level flag
This commit drops previous pre-release code that created --very-verbose
and --very-very-verbose flags, and instead uses the standard Python
logging levels (lowercased) with a --log-level flag. We omit the
"critical" log level since it is indistinguishable from "error".

Our new standard for output will be as follows:

- logger.error(): optionally used to give extra explanation when an
exception is raised
- logger.warning(): warn user about behavior that is most likely wrong
but not enough to cause an exception (default output level, so users
will see this for most models)
- logger.info(): high-level progress log; used to follow progress of the
model without seeing every detail
- logger.debug(): detailed diagnostic data (e.g., recommend improved
practices for input data files even if the current files are officially
acceptable)

We now issue deprecation warnings if users specify --verbose or --quiet.

This code creates a unique logger for each model, because different
models may be created during the life of the solve module and may use
different --log-run, --logs-dir and --log-level settings.

Code needs to be updated to send all reporting via the logging
mechanism. Then we may be able to drop the code that captures stdout and
stderr to the logfile. However, we will still need to find some way to
copy everything to the logfile, including tracebacks and messages from
other code (emitted via either print() or the root logger).