Don't report zonal electricity costs if zonal demand is not defined
In some experimental models, Switch is used as a general-purpose 
optimization framework without any electricity components (e.g., to 
prototype a gas network model). In this case, trying to report zonal 
electricity costs would cause an error, because those don't exist. We 
now test for this case and avoid that error.