Report location of model in callstack when using --debug
If there is an error and the --debug flag is used, the code now looks 
for a variable that contains a reference to a Pyomo model, and then 
reports the name of that variable and location in the callstack before 
running the post-mortem debugger. This may help users find a model to 
inspect and debug.