Count components as they are created when --log-level is 'info'
This intercepts stdout and converts timing messages from Pyomo into a 
stationary 'n/m components constructed' message. When writing to a file, 
the file handle is moved back so only the final message is written. If 
other messages are emitted along the way, they are not erased. This can 
interfere with the display a little during debugging.