Use 30% subsidy for distributed PV in 2020 and 2021
This branch uses an incorrect subsidy level for DistPV in 2020 and 2021 
(0.3 instead of 0.26 and 0.22). This should not affect results used for 
the RIST model because quantities of DistPV are fixed for that model and 
subsidies are not included in RIST inputs. However, the error is 
retained here to maximize reproducibility. For new work, it is 
recommended that you correct these subsidies (i.e., use a newer version 
of Switch).