Apply unserved_load at distribution node when available
This clarifies supply-demand balance and avoids infeasibility in
models that use switch_model.balancing.unserved_load with
switch_model.transmission.local_td (distribution-level balancing
in addition to zone-level balancing). This may slightly change
the unserved load and  total_cost reported in those models, due
to avoiding the need for local transmission and distribution to
meet the unserved portion of load.