Allow omission of any column in an optional input .csv file, even if the file is provided
It is confusing for users to specify that a table is optional, but then also have to
specify that individual columns are optional. This makes it possible for users just
to specify that the table is optional and leave it at that. Normally these params
would have a default value and get picked up that way, but in some cases they instead
define an indexing set that is used to identify special cases. It is confusing to have
to say that those particular columns are optional when the whole table is optional.