Improve code for progress messages during model construction
Pyomo takes a long time to convert an abstract model into an instance,
attaching all the data it has previously read from disk. This often
causes a multi-minute delay with no output from the software, which
can cause users to worry. Earlier code turned on Pyomo's report_timing
feature and then intercepted the messages from Pyomo and converted
them into progress messages, using backspaces to count up in the same
place on screen. That solution was brittle because it depends on
intercepting stdout (already complicated with our own logging and
Pyomo's logging going on) and identifying messages correctly (which can
change from one version of Pyomo or Python to another, e.g., it
appears Python switched from sending "\n" with the first write
call to sending it in a later call at some point). It also interfered
with debugging, since it threw away or diverted stdout, which made
pdb not display correctly if it was invoked during model construction
(which is common).

This new version overrides AbstractModel._initialize_component, which
Pyomo calls to initialize each component. It does the standard
behavior and then adds a simple progress message every 10% of the
way if --log-level is info or debug. This should be more reliable and
require less maintenance (it will just go quiet if Pyomo stops using
_initialize_component).