Rename gen_predetermined_cap to build_gen_predetermined in gen_build_predetermined.csv
This also renames gen_predetermined_storage_energy_mwh (new in this release, 2.0.7) to
build_gen_energy_predetermined. The new names should be more consistent with each other
and also be a little clearer that this is the amount of new capacity built in each year,
not the total amount of capacity in place that year.