Ensure every Param has domain and avoid using obj.keys() to initialize components
Both of these changes are needed for Pyomo 6+ compatibility. Pyomo issues a
deprecation warning for Params with default "Any" domain, due to future plans
to change this to "Reals". It also fails to clone models that use obj.keys()
to initialize components (probably not able to clone/pickle the iterator).