Use isinstance() to check whether a sum() of Pyomo components has any members
`build` is a Python variable created as the sum of several Pyomo components.
We previously checked whether it was empty via `build is 0`, but Python now
gives a warning about `is` with a literal, so we changed it to `build == 0`,
but that forces evaluation of an unconstructed component, which crashes
Pyomo. So instead we check `isinstance(build, int) and build == 0`. The
second part is not actually needed, but clarifies what is happening.