Move model configuration banner above define_components messages
To do this, I had to move the banner into SwitchAbstractModel.__init__
between the parameter parsing code and the define_components code
(or else split __init__ into multiple methods that have to be called
in the right order from switch_model.solve). To do that, I had to
make iteration_modules a part of the model instead of a temporary
variable in switch_model.solve, which seems reasonable anyway. This
commit also converts a number of print() calls into logger.info()
calls.