Raise Pyomo dependency to 5.7 and fix/tidy other dependencies
Pyomo versions before 5.7 had a bug where parameter data that was outside
the allowed domain would be treated as missing data. This could lead to
modeling errors if the data is replaced by default values, so we no longer
allow those versions of Pyomo. This means all previous users will need to
upgrade their version of Pyomo. This may also (correctly) cause errors to
be reported for models that previously seemed to run OK, but had invalid
data in optional columns.