add `switch info` with `--module-list` and `--module-arguments <modname>` flags
We are in the process of implementing a GUI wrapper for Switch, which will work
with a bundled copy of Switch that runs as a command-line application. The GUI
app needs a way to learn which modules are available and which arguments they
take. This provides that capability. Results will be pretty printed to stdout
as Python code by default (generally not very useful, but maybe for diagnosis),
or sent as json code if the --json flag is used (best for calling from other
programs).

The `--module-list` flag gives a list of all available modules in the search
path. In the future this will be updated to return them in the order they can
appear in modules.txt.

The `--module-arguments <modname>` command reports all the command-line
arguments that can be accepted by module <modname>.