Implement --retire {early, mid, late} flag
This flag controls whether to retire projects at the start of the period
when they reach end-of-life ('early') (i.e., only run if they survive to
the end of the period), or retire them if they survive past the middle of
the period ('mid'), or extend operation to the end of the period when
they reach end-of-life ('late'). Late is the default and matches
previous behavior by Switch. Early and mid match some other models'
behavior.