Add generator retrofit capability (switch_model.generators.extensions.retrofit)
The switch_model.generators.extensions.retrofit module makes it possible to
define retrofits for existing generators. These are done by defining a new
generation project that can replace a previously built one (i.e., it performs
like the original generator plus the retrofit) and adding columns to
`gen_retrofits.csv` showing all allowed combinations of base projects that can
be replaced by retrofit projects. In each row, `base_gen_project` shows the
name of the original (base) project and `retrofit_gen_project` shows the name
of a retrofit project that can replace it.

Retrofit projects will only be built if the base project has also been built
in the same or an earlier period. When a base project is retrofitted, the base
project is suspended (via SuspendGen) and the retrofit version is built and
operated instead. In addition, retrofit projects are automatically suspended
at the end of life of the base project. (To enable these behaviors,
gen_can_retire_early or gen_can_suspend must be set to True or 1 in
gen_info.csv for both the base project and the retrofit version.)

Because of this framing, retrofitted projects will have capital expenditure
equal to the capital recovery for the base project plus capital recovery for
the retrofit project. So gen_overnight_cost for the retrofit project should be
set equal to the cost of the retrofit work, not the combined project. However,
fixed and variable O&M will no longer be collected for the base project, so
O&M cost inputs for the retrofit project should be the ones that apply for the
total retrofitted project.

Capital recovery for the retrofitted project will be amortized over the
remaining life of the base project that it replaces, which may cause faster
capital recovery than would otherwise be expected for these assets.