Add tp_date parameter to identify which date each timepoint falls on
This is used by modules that enforce constraints between hours of the
same date (e.g., hydrogen production or simple demand response). It
defaults to be the same as the timeseries, which works fine if each
timeseries is one day long. But if models use multi-day timeseries,
they should specify a code for the corresponding date for each timepoint
in the tp_date column of inputs/timepoints.csv.