Track hydro reservoir levels by timeseries instead of period
The hydro_system module now tracks reservoir levels over the course of each
timeseries instead of each period. Previously the hydro_system module tracked
levels of reservoirs across all the timepoints in a period, as if each
timeseries was linked to the one after it. This was an error, and in models
with multiple timeseries with different weights, Switch could retain water
during low-weight timeseries, then release it during high-weight timeseries,
effectively producing free energy.

The new approach matches the general principle in Switch that each timeseries is
independent and it must be possible to to repeat each timeseries many times in a
row if necessary. Although this is more correct than the previous approach, it
is also more conservative: hydro networks must reach the same level (or a
prespecified level) at the start and end of each timeseries, instead of the
start and end of each period.

Users should now set res_initial_vol and res_final_vol for each timeseries in
reservoir_ts_data.csv instead of setting initial_res_vol and final_res_vol for
each period in reservoirs.csv. The standard upgrade script will automatically
make this change if needed. It will also interpolate between the previous
full-period reservoir volumes to set the reservoir volumes at the start and end
of each timeseries.

For models with a single, long timeseries for each period (the most common use
case for this module), this change won't make a difference. However, for models
with multiple timeseries in each period with different weights, this update
may change the model results.

In addition, res_initial_vol and res_final_vol are now optional. If
res_initial_vol is set but not res_final_vol, Switch will set the final level
equal to the initial level. If neither is set, Switch will choose an optimal
initial volume for each timeseries and also return to that level at the end of
the timeseries. (For most models, it is best not to set res_final_vol, so Switch
will return the reservoir to the starting level. This makes it possible to have
an arbitrary number of repetitions of each timeseries.)