Improve rules for refinery cogen plants in hawaii.oahu_plants
Prior to this commit, there was special code in the oahu_plants module to
check that the refinery cogen plants existed with the expected names
(hadn't been renamed) and to require they shut down when the refineries
are closed and that they never burn biofuel. The first check could crash
models that start in 2045 or later, because the plants will be retired
by then.

This commit makes the check more generic: it verifies that all cogen
plants run MSW or LSFO (not something new that might throw off the
behavior) and also adds code to only allow a single fuel if specified
(not auto-switch to any lighter fuel). The refineries are now marked as
"LSFO only" in the input file, so they will come in with "LSFO" as
gen_energy_source and pass the first check. This also makes it unnecessary
to force the cogen plants not to use biodiesel in the oahu_plants code,
because that is now specified in the existing plants input file. So the
no-biodiesel code has been removed from oahu_plants.