Allow later versions of Pyomo and Python
Previously Switch 2.0.9 required Pyomo <= 6.7.2, which is incompatible
with numpy >= 2.0, which would automatically be installed as part of the
pandas dependency after June 2024. This release updates metadata to
allow later versions of Pyomo that will work with numpy > 2.0. Note that
it is still possible for users to create a non-working environment by
installing Pyomo < 6.8.0 with numpy >= 2.0, but we assume most installations
will either use old versions of Pyomo with old versions of Numpy
or new versions of Pyomo with new versions of Numpy.